Rename instances of Compiler to CodeGenerator, where appropriate
