+++ Source/DafnyCore/Backends/CSharp/CsharpSynthesizer.cs
@@ -179,7 +179,7 @@ public class CsharpSynthesizer {
     var methodName = method.GetCompileName(Options);
 
     if (((Function)method).Ens.Count != 0) {
-      codeGenerator.Error(CompilerErrors.ErrorId.c_possibly_unsatisfied_postconditions, lastSynthesizedMethod.tok,
+      codeGenerator.Error(GeneratorErrors.ErrorId.c_possibly_unsatisfied_postconditions, lastSynthesizedMethod.tok,
         "Post-conditions on function {0} might " +
         "be unsatisfied when synthesizing code " +
         "for method {1}",
@@ -223,7 +223,7 @@ public class CsharpSynthesizer {
       var obj = ((IdentifierExpr)exprDotName.Lhs.Resolved).Var;
       var field = ((MemberSelectExpr)exprDotName.Resolved).Member;
       var fieldName = field.GetCompileName(Options);
-      codeGenerator.Error(CompilerErrors.ErrorId.c_stubbing_fields_not_recommended, lastSynthesizedMethod.tok,
+      codeGenerator.Error(GeneratorErrors.ErrorId.c_stubbing_fields_not_recommended, lastSynthesizedMethod.tok,
         "Stubbing fields is not recommended (field {0} of object {1} inside method {2})",
         ErrorWriter, fieldName, obj.Name, lastSynthesizedMethod.Name);
       var tmpId = codeGenerator.idGenerator.FreshId("tmp");
