+++ Source/DafnyCore/Backends/CoverageInstrumenter.cs
