+++ Source/DafnyCore/Backends/DafnyExecutableBackend.cs
@@ -11,7 +11,7 @@ namespace Microsoft.Dafny.Compilers;
 
 public abstract class DafnyExecutableBackend : ExecutableBackend {
 
-  protected DafnyWrittenCompiler dafnyCompiler;
+  protected DafnyWrittenCodeGenerator DafnyCodeGenerator;
 
   protected DafnyExecutableBackend(DafnyOptions options) : base(options) {
   }
@@ -21,11 +21,11 @@ public abstract class DafnyExecutableBackend : ExecutableBackend {
     return new DafnyCodeGenerator(Options, Reporter);
   }
 
-  protected abstract DafnyWrittenCompiler CreateDafnyWrittenCompiler();
+  protected abstract DafnyWrittenCodeGenerator CreateDafnyWrittenCompiler();
 
   public override void OnPreCompile(ErrorReporter reporter, ReadOnlyCollection<string> otherFileNames) {
     base.OnPreCompile(reporter, otherFileNames);
-    dafnyCompiler = CreateDafnyWrittenCompiler();
+    DafnyCodeGenerator = CreateDafnyWrittenCompiler();
   }
 
   public override void Compile(Program dafnyProgram, ConcreteSyntaxTree output) {
@@ -34,12 +34,12 @@ public abstract class DafnyExecutableBackend : ExecutableBackend {
     ((DafnyCodeGenerator)codeGenerator).Start();
     codeGenerator.Compile(dafnyProgram, new ConcreteSyntaxTree());
     var dast = ((DafnyCodeGenerator)codeGenerator).Build();
-    var o = dafnyCompiler.Compile((Sequence<DAST.Module>)Sequence<DAST.Module>.FromArray(dast.ToArray()));
+    var o = DafnyCodeGenerator.Compile((Sequence<DAST.Module>)Sequence<DAST.Module>.FromArray(dast.ToArray()));
     output.Write(o.ToVerbatimString(false));
   }
 
   public override void EmitCallToMain(Method mainMethod, string baseName, ConcreteSyntaxTree output) {
-    var o = dafnyCompiler.EmitCallToMain(mainMethod.FullSanitizedName);
+    var o = DafnyCodeGenerator.EmitCallToMain(mainMethod.FullSanitizedName);
     output.Write(o.ToVerbatimString(false));
   }
