+++ Source/DafnyCore/Backends/DafnyWrittenCodeGenerator.cs
@@ -3,7 +3,7 @@ using Dafny;
 
 namespace Microsoft.Dafny.Compilers {
 
-  public abstract class DafnyWrittenCompiler {
+  public abstract class DafnyWrittenCodeGenerator {
 
     public abstract ISequence<Rune> Compile(Sequence<DAST.Module> program);
