+++ Source/DafnyCore/Backends/DatatypeWrapperEraser.cs
