+++ Source/DafnyCore/Backends/ExecutableBackend.cs
@@ -1,17 +1,12 @@
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
-using System.CommandLine;
-using System.Data.SqlTypes;
 using System.Diagnostics;
 using System.Diagnostics.Contracts;
 using System.IO;
 using System.Linq;
-using System.Runtime.InteropServices;
-using System.Text;
 using System.Threading.Tasks;
 using Microsoft.Dafny.Plugins;
-using Microsoft.Win32;
 
 namespace Microsoft.Dafny.Compilers;
