+++ Source/DafnyCore/Backends/JavaScript/JavaScriptCodeGenerator.cs
